+++ backend/src/main/resources/application-local.yml
@@ -14,9 +14,10 @@ spring:
     activate:
       on-profile: local
   jpa:
+    generate-ddl: true
     open-in-view: false
     hibernate:
-      ddl-auto: create
+      ddl-auto: update
     properties:
       hibernate:
         dialect: org.hibernate.dialect.MySQL57Dialect
@@ -24,7 +25,7 @@ spring:
   datasource:
     url: jdbc:mysql://localhost:3306/ternoko?serverTimezone=Asia/Seoul
     username: root
-    password:
+    password: 1234
 
 slack:
   clientId: '3756998338916.3821665111344'
