+++ backend/src/test/java/com/woowacourse/ternoko/comment/application/CommentServiceTest.java
@@ -5,8 +5,10 @@ import static com.woowacourse.ternoko.common.exception.ExceptionType.INTERVIEW_N
 import static com.woowacourse.ternoko.common.exception.ExceptionType.INVALID_COMMENT_INTERVIEW_ID;
 import static com.woowacourse.ternoko.common.exception.ExceptionType.INVALID_COMMENT_MEMBER_ID;
 import static com.woowacourse.ternoko.common.exception.ExceptionType.INVALID_INTERVIEW_MEMBER_ID;
+import static com.woowacourse.ternoko.common.exception.ExceptionType.INVALID_STATUS_CREATE_COMMENT;
 import static com.woowacourse.ternoko.common.exception.ExceptionType.INVALID_STATUS_FIND_COMMENT;
 import static com.woowacourse.ternoko.fixture.MemberFixture.COACH1;
+import static com.woowacourse.ternoko.fixture.MemberFixture.COACH2;
 import static com.woowacourse.ternoko.fixture.MemberFixture.CREW1;
 import static com.woowacourse.ternoko.fixture.MemberFixture.CREW2;
 import static org.assertj.core.api.Assertions.assertThat;
@@ -18,10 +20,13 @@ import com.woowacourse.ternoko.comment.dto.CommentsResponse;
 import com.woowacourse.ternoko.comment.exception.CommentNotFoundException;
 import com.woowacourse.ternoko.comment.exception.InvalidCommentInterviewIdException;
 import com.woowacourse.ternoko.comment.exception.InvalidCommentMemberIdException;
+import com.woowacourse.ternoko.comment.exception.InvalidStatusCreateCommentException;
 import com.woowacourse.ternoko.comment.exception.InvalidStatusFindCommentException;
 import com.woowacourse.ternoko.common.exception.MemberNotFoundException;
+import com.woowacourse.ternoko.domain.InterviewStatusType;
 import com.woowacourse.ternoko.domain.MemberType;
 import com.woowacourse.ternoko.interview.application.InterviewService;
+import com.woowacourse.ternoko.interview.dto.InterviewResponse;
 import com.woowacourse.ternoko.interview.exception.InterviewNotFoundException;
 import java.util.stream.Collectors;
 import org.junit.jupiter.api.DisplayName;
@@ -36,7 +41,130 @@ import org.springframework.transaction.annotation.Transactional;
 public class CommentServiceTest {
 
     @Autowired
+    private CommentService commentService;
+
+    @Autowired
     private InterviewService interviewService;
+
+    @Test
+    @DisplayName("[크루] 면담 종료 시 면담에 대한 한마디를 작성하면 면담이 완료된다.")
+    void createCommentByCrew() {
+        // given
+        final Long interviewId = 1L;
+        final CommentRequest commentRequest = new CommentRequest("너무나도 유익한 시간이었습니다. 감사합니다.");
+        // when
+        Long commentId = commentService.create(CREW1.getId(), interviewId, commentRequest);
+        // then
+        InterviewResponse interviewResponseById = interviewService.findInterviewResponseById(interviewId);
+        assertThat(commentId).isNotNull();
+        assertThat(interviewResponseById.getStatus()).isEqualTo(InterviewStatusType.CREW_COMPLETED);
+    }
+
+    @Test
+    @DisplayName("[코치] 면담 종료 시 면담에 대한 한마디를 작성하면 면담이 완료된다.")
+    void createCommentByCoach() {
+        // given
+        final Long interviewId = 1L;
+        final CommentRequest commentRequest = new CommentRequest("너무나도 유익한 시간이었습니다. 감사합니다.");
+        // when
+        Long commentId = commentService.create(COACH1.getId(), interviewId, commentRequest);
+        // then
+        InterviewResponse interviewResponseById = interviewService.findInterviewResponseById(interviewId);
+        assertThat(commentId).isNotNull();
+        assertThat(interviewResponseById.getStatus()).isEqualTo(InterviewStatusType.COACH_COMPLETED);
+    }
+
+    @Test
+    @DisplayName("면담에 관련되지 않은 회원이 코멘트를 작성하면 예외를 반환한다.")
+    void createComment_InvalidMember() {
+        // given
+        final Long interviewId = 1L;
+        final CommentRequest commentRequest = new CommentRequest("너무나도 유익한 시간이었습니다. 감사합니다.");
+        // when & then
+        assertThatThrownBy(() -> commentService.create(COACH2.getId(), interviewId, commentRequest))
+                .isInstanceOf(MemberNotFoundException.class)
+                .hasMessage(INVALID_INTERVIEW_MEMBER_ID.getMessage());
+    }
+
+    @Test
+    @DisplayName("존재하지 않는 면담에 코멘트를 작성하면 예외를 반환한다.")
+    void createComment_InvalidInterview() {
+        // given
+        final Long interviewId = 10L;
+        final CommentRequest commentRequest = new CommentRequest("너무나도 유익한 시간이었습니다. 감사합니다.");
+        // when & then
+        assertThatThrownBy(() -> commentService.create(COACH1.getId(), interviewId, commentRequest))
+                .isInstanceOf(InterviewNotFoundException.class)
+                .hasMessage(interviewId + INTERVIEW_NOT_FOUND.getMessage());
+    }
+
+    @Test
+    @DisplayName("[크루] 2번 이상 코멘트를 남기면 예외를 발생한다.")
+    void createCommentByCrew_InvalidStatus_Twice() {
+        // given
+        final Long interviewId = 1L;
+        final CommentRequest commentRequest = new CommentRequest("너무나도 유익한 시간이었습니다. 감사합니다.");
+        // when
+        commentService.create(CREW1.getId(), interviewId, commentRequest);
+        assertThatThrownBy(() -> commentService.create(CREW1.getId(), interviewId, commentRequest))
+                .isInstanceOf(InvalidStatusCreateCommentException.class)
+                .hasMessage(INVALID_STATUS_CREATE_COMMENT.getMessage());
+    }
+
+    @Test
+    @DisplayName("[코치] COMMENT, CREW_COMPLETE 이외의 상태에서 코멘트를 남기면 예외를 발생한다.")
+    void createCommentByCoach_InvalidStatus_Twice() {
+        // given
+        final Long interviewId = 1L;
+        final CommentRequest commentRequest = new CommentRequest("너무나도 유익한 시간이었습니다. 감사합니다.");
+        // when & then
+        commentService.create(COACH1.getId(), interviewId, commentRequest);
+        assertThatThrownBy(() -> commentService.create(COACH1.getId(), interviewId, commentRequest))
+                .isInstanceOf(InvalidStatusCreateCommentException.class)
+                .hasMessage(INVALID_STATUS_CREATE_COMMENT.getMessage());
+    }
+
+    @Test
+    @DisplayName("[코치] COMMENT, CREW_COMPLETE 이외의 상태에서 코멘트를 남기면 예외를 발생한다.")
+    void createCommentByCoach_InvalidStatus_OtherStatus() {
+        // given
+        final Long interviewId = 2L; // FIXED인 인터뷰
+        final CommentRequest commentRequest = new CommentRequest("너무나도 유익한 시간이었습니다. 감사합니다.");
+        // when & then
+        assertThatThrownBy(() -> commentService.create(COACH1.getId(), interviewId, commentRequest))
+                .isInstanceOf(InvalidStatusCreateCommentException.class)
+                .hasMessage(INVALID_STATUS_CREATE_COMMENT.getMessage());
+    }
+
+    @Test
+    @DisplayName("[크루] 코치가 한마디 작성 후 크루가 작성한다면 면담 상태는 COMPLETE가 된다.")
+    void createComment_Complete_ByCrew() {
+        // given
+        final Long interviewId = 1L;
+        final CommentRequest commentRequest = new CommentRequest("너무나도 유익한 시간이었습니다. 감사합니다.");
+        commentService.create(COACH1.getId(), interviewId, commentRequest);
+        // when
+        Long commentId = commentService.create(CREW1.getId(), interviewId, commentRequest);
+        // then
+        InterviewResponse interviewResponseById = interviewService.findInterviewResponseById(interviewId);
+        assertThat(commentId).isNotNull();
+        assertThat(interviewResponseById.getStatus()).isEqualTo(InterviewStatusType.COMPLETE);
+    }
+
+    @Test
+    @DisplayName("[코치] 크루가 한마디 작성 후 코치가 작성한다면 면담 상태는 COMPLETE가 된다.")
+    void createComment_Complete_ByCoach() {
+        // given
+        final Long interviewId = 1L;
+        final CommentRequest commentRequest = new CommentRequest("너무나도 유익한 시간이었습니다. 감사합니다.");
+        commentService.create(CREW1.getId(), interviewId, commentRequest);
+        // when
+        Long commentId = commentService.create(COACH1.getId(), interviewId, commentRequest);
+        // then
+        InterviewResponse interviewResponseById = interviewService.findInterviewResponseById(interviewId);
+        assertThat(commentId).isNotNull();
+        assertThat(interviewResponseById.getStatus()).isEqualTo(InterviewStatusType.COMPLETE);
+    }
 
     @Test
     @DisplayName("[크루] CREW_COMMENT, COMPLETE 상태에서 조회가 가능하다.")
@@ -44,9 +172,9 @@ public class CommentServiceTest {
         // given
         final Long interviewId = 1L;
         final CommentRequest commentRequest = new CommentRequest("너무나도 유익한 시간이었습니다. 감사합니다.");
-        interviewService.createComment(CREW1.getId(), interviewId, commentRequest);
+        commentService.create(CREW1.getId(), interviewId, commentRequest);
         // when
-        final CommentsResponse commentsResponse = interviewService.findComments(CREW1.getId(), interviewId);
+        final CommentsResponse commentsResponse = commentService.findComments(CREW1.getId(), interviewId);
         // then
         assertThat(commentsResponse.getComments().size()).isEqualTo(1);
         assertThat(commentsResponse.getComments().get(0).getComment()).isEqualTo("너무나도 유익한 시간이었습니다. 감사합니다.");
@@ -59,9 +187,9 @@ public class CommentServiceTest {
         // given
         final Long interviewId = 1L;
         final CommentRequest commentRequest = new CommentRequest("너무나도 유익한 시간이었습니다. 감사합니다.");
-        interviewService.createComment(COACH1.getId(), interviewId, commentRequest);
+        commentService.create(COACH1.getId(), interviewId, commentRequest);
         // when
-        final CommentsResponse commentsResponse = interviewService.findComments(COACH1.getId(), interviewId);
+        final CommentsResponse commentsResponse = commentService.findComments(COACH1.getId(), interviewId);
         // then
         assertThat(commentsResponse.getComments().size()).isEqualTo(1);
         assertThat(commentsResponse.getComments().get(0).getComment()).isEqualTo("너무나도 유익한 시간이었습니다. 감사합니다.");
@@ -75,10 +203,10 @@ public class CommentServiceTest {
         final Long interviewId = 1L;
         final CommentRequest coachCommentRequest = new CommentRequest("재밌는 시간 고마워요. 고민이 해결되기를 바랄게요.");
         final CommentRequest crewCommentRequest = new CommentRequest("너무나도 유익한 시간이었습니다. 감사합니다.");
-        interviewService.createComment(COACH1.getId(), interviewId, coachCommentRequest);
-        interviewService.createComment(CREW1.getId(), interviewId, crewCommentRequest);
+        commentService.create(COACH1.getId(), interviewId, coachCommentRequest);
+        commentService.create(CREW1.getId(), interviewId, crewCommentRequest);
         // when
-        final CommentsResponse commentsResponse = interviewService.findComments(COACH1.getId(), interviewId);
+        final CommentsResponse commentsResponse = commentService.findComments(COACH1.getId(), interviewId);
         // then
         assertThat(commentsResponse.getComments().size()).isEqualTo(2);
         assertThat(commentsResponse.getComments().stream()
@@ -97,9 +225,9 @@ public class CommentServiceTest {
         // given
         final Long interviewId = 1L;
         final CommentRequest commentRequest = new CommentRequest("재밌는 시간 고마워요. 고민이 해결되기를 바랄게요.");
-        interviewService.createComment(COACH1.getId(), interviewId, commentRequest);
+        commentService.create(COACH1.getId(), interviewId, commentRequest);
         // when &  then
-        assertThatThrownBy(() -> interviewService.findComments(CREW1.getId(), interviewId))
+        assertThatThrownBy(() -> commentService.findComments(CREW1.getId(), interviewId))
                 .isInstanceOf(InvalidStatusFindCommentException.class)
                 .hasMessage(INVALID_STATUS_FIND_COMMENT.getMessage());
     }
@@ -110,9 +238,9 @@ public class CommentServiceTest {
         // given
         final Long interviewId = 1L;
         final CommentRequest commentRequest = new CommentRequest("너무나도 유익한 시간이었습니다. 감사합니다.");
-        interviewService.createComment(CREW1.getId(), interviewId, commentRequest);
+        commentService.create(CREW1.getId(), interviewId, commentRequest);
         // when &  then
-        assertThatThrownBy(() -> interviewService.findComments(COACH1.getId(), interviewId))
+        assertThatThrownBy(() -> commentService.findComments(COACH1.getId(), interviewId))
                 .isInstanceOf(InvalidStatusFindCommentException.class)
                 .hasMessage(INVALID_STATUS_FIND_COMMENT.getMessage());
     }
@@ -123,7 +251,7 @@ public class CommentServiceTest {
         // given
         final Long interviewId = 1L;
         // when &  then
-        assertThatThrownBy(() -> interviewService.findComments(COACH1.getId(), interviewId))
+        assertThatThrownBy(() -> commentService.findComments(COACH1.getId(), interviewId))
                 .isInstanceOf(InvalidStatusFindCommentException.class)
                 .hasMessage(INVALID_STATUS_FIND_COMMENT.getMessage());
     }
@@ -134,7 +262,7 @@ public class CommentServiceTest {
         // given
         final Long interviewId = 10L;
         // when & then
-        assertThatThrownBy(() -> interviewService.findComments(COACH1.getId(), interviewId))
+        assertThatThrownBy(() -> commentService.findComments(COACH1.getId(), interviewId))
                 .isInstanceOf(InterviewNotFoundException.class)
                 .hasMessage(interviewId + INTERVIEW_NOT_FOUND.getMessage());
     }
@@ -145,9 +273,9 @@ public class CommentServiceTest {
         // given
         final Long interviewId = 1L;
         final CommentRequest commentRequest = new CommentRequest("너무나도 유익한 시간이었습니다. 감사합니다.");
-        interviewService.createComment(CREW1.getId(), interviewId, commentRequest);
+        commentService.create(CREW1.getId(), interviewId, commentRequest);
         // when & then
-        assertThatThrownBy(() -> interviewService.createComment(CREW2.getId(), interviewId, commentRequest))
+        assertThatThrownBy(() -> commentService.create(CREW2.getId(), interviewId, commentRequest))
                 .isInstanceOf(MemberNotFoundException.class)
                 .hasMessage(INVALID_INTERVIEW_MEMBER_ID.getMessage());
     }
@@ -158,11 +286,11 @@ public class CommentServiceTest {
         // given
         final Long interviewId = 1L;
         final CommentRequest commentRequest = new CommentRequest("너무나도 유익한 시간이었습니다. 감사합니다.");
-        Long commentId = interviewService.createComment(CREW1.getId(), interviewId, commentRequest);
+        Long commentId = commentService.create(CREW1.getId(), interviewId, commentRequest);
         // when
         final CommentRequest updateCommentRequest = new CommentRequest("이 말을 빼먹었어요~");
-        interviewService.updateComment(CREW1.getId(), interviewId, commentId, updateCommentRequest);
-        final CommentsResponse commentsResponse = interviewService.findComments(CREW1.getId(), interviewId);
+        commentService.update(CREW1.getId(), interviewId, commentId, updateCommentRequest);
+        final CommentsResponse commentsResponse = commentService.findComments(CREW1.getId(), interviewId);
         // then
         assertThat(commentsResponse.getComments().size()).isEqualTo(1);
         assertThat(commentsResponse.getComments().get(0).getComment()).isEqualTo("이 말을 빼먹었어요~");
@@ -175,10 +303,10 @@ public class CommentServiceTest {
         // given
         final Long interviewId = 1L;
         final CommentRequest commentRequest = new CommentRequest("너무나도 유익한 시간이었습니다. 감사합니다.");
-        Long commentId = interviewService.createComment(CREW1.getId(), interviewId, commentRequest);
+        Long commentId = commentService.create(CREW1.getId(), interviewId, commentRequest);
         // when & then
         final CommentRequest updateCommentRequest = new CommentRequest("이 말을 빼먹었어요~");
-        assertThatThrownBy(() -> interviewService.updateComment(CREW1.getId(), interviewId, 10L, updateCommentRequest))
+        assertThatThrownBy(() -> commentService.update(CREW1.getId(), interviewId, 10L, updateCommentRequest))
                 .isInstanceOf(CommentNotFoundException.class)
                 .hasMessage(10L + COMMENT_NOT_FOUND.getMessage());
     }
@@ -190,13 +318,13 @@ public class CommentServiceTest {
         final Long firstInterviewId = 1L;
         final Long secondInterviewId = 3L;
         final CommentRequest firstRequest = new CommentRequest("너무나도 유익한 시간이었습니다. 감사합니다.");
-        Long firstCommentId = interviewService.createComment(CREW1.getId(), firstInterviewId, firstRequest);
+        Long firstCommentId = commentService.create(CREW1.getId(), firstInterviewId, firstRequest);
 
         final CommentRequest secondRequest = new CommentRequest("너무나도 유익한 시간이었습니다. 감사합니다.");
-        Long secondCommentId = interviewService.createComment(CREW1.getId(), secondInterviewId, secondRequest);
+        Long secondCommentId = commentService.create(CREW1.getId(), secondInterviewId, secondRequest);
         // when & then
         final CommentRequest updateCommentRequest = new CommentRequest("이 말을 빼먹었어요~");
-        assertThatThrownBy(() -> interviewService.updateComment(CREW1.getId(), firstInterviewId, secondCommentId,
+        assertThatThrownBy(() -> commentService.update(CREW1.getId(), firstInterviewId, secondCommentId,
                 updateCommentRequest))
                 .isInstanceOf(InvalidCommentInterviewIdException.class)
                 .hasMessage(INVALID_COMMENT_INTERVIEW_ID.getMessage());
@@ -208,13 +336,13 @@ public class CommentServiceTest {
         // given
         final Long interviewId = 1L;
         final CommentRequest crewCommentRequest = new CommentRequest("너무나도 유익한 시간이었습니다. 감사합니다.");
-        interviewService.createComment(CREW1.getId(), interviewId, crewCommentRequest);
+        commentService.create(CREW1.getId(), interviewId, crewCommentRequest);
         final CommentRequest coachCommentRequest = new CommentRequest("즐거운 면담이었습니다. 오늘도 즐거운 하루 되세요.");
-        Long coachCommentId = interviewService.createComment(COACH1.getId(), interviewId, coachCommentRequest);
+        Long coachCommentId = commentService.create(COACH1.getId(), interviewId, coachCommentRequest);
         // when & then
         final CommentRequest updateCommentRequest = new CommentRequest("이 말을 빼먹었어요~");
         assertThatThrownBy(
-                () -> interviewService.updateComment(CREW1.getId(), interviewId, coachCommentId, updateCommentRequest))
+                () -> commentService.update(CREW1.getId(), interviewId, coachCommentId, updateCommentRequest))
                 .isInstanceOf(InvalidCommentMemberIdException.class)
                 .hasMessage(INVALID_COMMENT_MEMBER_ID.getMessage());
     }
